backend/worker/requirements.txt
# =====================================================================
# Sprint 1 — Dummy Worker Dependencies
# =====================================================================
kafka-python-ng
pymongo==4.7.3
python-dotenv==1.0.1

# =====================================================================
# Sprint 2 & 3 — AI Worker Dependencies
# =====================================================================

# OpenAI client
openai>=1.50.0
httpx==0.27.2

# LangChain core + OpenAI integration
langchain==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5

# Tavily web search
tavily-python==0.3.3
